+++ relay-server/src/envelope.rs
@@ -106,6 +106,8 @@ pub enum ItemType {
     ProfilingSession,
     /// Profiling data
     ProfilingTrace,
+    /// Profile event payload encoded in JSON
+    Profile,
 }
 
 impl ItemType {
@@ -139,6 +141,7 @@ impl fmt::Display for ItemType {
             Self::ClientReport => write!(f, "client report"),
             Self::ProfilingSession => write!(f, "profiling session"),
             Self::ProfilingTrace => write!(f, "profiling trace"),
+            Self::Profile => write!(f, "profile"),
         }
     }
 }
@@ -584,7 +587,8 @@ impl Item {
             | ItemType::MetricBuckets
             | ItemType::ClientReport
             | ItemType::ProfilingSession
-            | ItemType::ProfilingTrace => false,
+            | ItemType::ProfilingTrace
+            | ItemType::Profile => false,
         }
     }
 
@@ -608,6 +612,7 @@ impl Item {
             ItemType::ClientReport => false,
             ItemType::ProfilingSession => false,
             ItemType::ProfilingTrace => false,
+            ItemType::Profile => false,
         }
     }
 }